refac: Rename lockfiles.rs -> tools.rs
