+++ src/tools/mod.rs
@@ -1,5 +1,5 @@
 pub mod install;
 pub mod list;
-pub mod lockfiles;
 mod pixi_config;
+pub mod tools;
 pub mod uninstall;
